Added the first token catagory and major reformatting
after redoing the page format to be ore flexible, added a token catagory that is fully working outide of the tokens actual effects. It tracks progress as well as allows for the purchase of tokens. Will copy and paste this to create the other catagories. Added a simple outline for the catagories to the notes file.